add：DataClassのItemがViewModel内にあったため、独立させました
